requirements.txt
gradio==4.44.1
gradio_image_prompter
diffusers
transformers==4.57.0
pydantic==2.10.6
einops
torch-cluster
huggingface_hub
opencv-python
trimesh
omegaconf
scikit-image
numpy
peft
pytorch-lightning
open3d
pymeshlab==2022.2.post4
jaxtyping
numpy==1.26.0
ninja
cython
pip install gltflib
gltlib